Minor long deserialization improvements.
Since long is not referenced from other parts, inline it's loop into main function.
Changed if range else into optional if condition.
This seems to help a little since it can be assumed that if path will never be taked and no checks need to be done in that case.

While int cound benefit from same optimizations, it's referenced from other parts (exponent parsing) so leave it as is.